LF-5211 Add a file for uploadFarmNoteImage middleware
